Remove mock data
