--- src/mock/filter.js
@@ -1,14 +1,0 @@
-export const filters = [
-  {
-    name: `everything`,
-    checked: true
-  },
-  {
-    name: `future`,
-    checked: false
-  },
-  {
-    name: `past`,
-    checked: false
-  }
-];
