--- src/mock/menu.js
@@ -1,15 +1,0 @@
-export const MenuItem = {
-  TABLE: `Table`,
-  STATS: `Stats`
-};
-
-export const menuItems = [
-  {
-    name: MenuItem.TABLE,
-    active: true
-  },
-  {
-    name: MenuItem.STATS,
-    active: false
-  }
-];
